refactor: move Action enum
