+++ views/components/NodeListItem.test.tsx
@@ -11,8 +11,9 @@ import userEvent from '@testing-library/user-event';
 
 import { PREVIEW_PATH, PREVIEW_TYPE, REST_ENDPOINT, ROOTS } from '../../constants';
 import { populateFile, populateFolder, populateNode, populateUser } from '../../mocks/mockUtils';
+import { Action } from '../../types/common';
 import { NodeType, User } from '../../types/graphql/types';
-import { Action, getPermittedHoverBarActions } from '../../utils/ActionsFactory';
+import { getPermittedHoverBarActions } from '../../utils/ActionsFactory';
 import { render } from '../../utils/testUtils';
 import { formatDate, humanFileSize } from '../../utils/utils';
 import * as moduleUtils from '../../utils/utils';
