+++ views/components/UploadListItem.tsx
@@ -14,9 +14,9 @@ import { useNavigation } from '../../../hooks/useNavigation';
 import { LIST_ITEM_HEIGHT } from '../../constants';
 import { Breadcrumbs } from '../../design_system_fork/Breadcrumbs';
 import { useUpload } from '../../hooks/useUpload';
-import { UploadStatus } from '../../types/common';
+import { Action, UploadStatus } from '../../types/common';
 import { Maybe, Node } from '../../types/graphql/types';
-import { Action, buildActionItems } from '../../utils/ActionsFactory';
+import { buildActionItems } from '../../utils/ActionsFactory';
 import { buildCrumbs, humanFileSize, scrollToNodeItem } from '../../utils/utils';
 import { ContextualMenu } from './ContextualMenu';
 import { NodeAvatarIcon } from './NodeAvatarIcon';
